requirements.txt
pydantic~=2.6.1
asyncpg~=0.29.0
python-dotenv~=1.0.1
email-validator~=2.1.0.post1
alembic~=1.13.1